Improvement list comments view and a lot of changes in JS and serializers
