+++ src/laboratory/templates/laboratory/laboratoryroom_list.html
@@ -139,6 +139,7 @@ const view_search = false;
 <script src="{% static 'js/shelfobjectedit.js' %}?v={% get_organilab_version %}"></script>
 <script src="{% static 'js/reservationmodal.js' %}?v={% get_organilab_version %}"></script>
 
+<script type="text/javascript" src="{% static 'laboratory/js/base_modal_management.js'%}?v={% get_organilab_version %}"></script>
 <script type="text/javascript" src="{% static 'laboratory/js/laboratory.js'%}?v={% get_organilab_version %}"></script>
